+++ admin-app/src/components/style.css
@@ -1,0 +1,30 @@
+.headerWrap{
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  padding: 10px 50px;
+}
+
+.headerWrap div{
+  display: flex;
+  align-items: center;
+}
+
+.headerWrap h1{
+  font-size: 35px;
+}
+
+nav{
+  margin-left: 80px;
+}
+
+nav ul li{
+  display: inline-block;
+  font-size: 20px;
+  margin: 0 20px;
+}
+
+p{
+  font-size: 18px;
+}
