+++ admin-app/src/vite-env.d.ts
