+++ admin-app/vite.config.ts
